+++ compiler/src/Language/Mimsa/Typechecker/Elaborate.hs
@@ -4,9 +4,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 
 module Language.Mimsa.Typechecker.Elaborate
-  ( runElabM,
-    elab,
-    elabAndSubst,
+  ( elab,
     recoverAnn,
     TypedAnnotation,
     getTypeFromAnn,
@@ -15,11 +13,10 @@ where
 
 import Control.Monad.Except
 import Control.Monad.Reader
-import Control.Monad.State (State, runState)
+import Control.Monad.State (State)
 import Control.Monad.Writer
 import Data.Foldable
 import qualified Data.List.NonEmpty as NE
-import Data.Map (Map)
 import qualified Data.Map as M
 import Data.Maybe (listToMaybe)
 import qualified Data.Set as S
@@ -37,37 +34,6 @@ import Language.Mimsa.Types.Swaps
 import Language.Mimsa.Types.Typechecker
 
 type ElabM = ExceptT TypeError (WriterT [Constraint] (ReaderT Swaps (State TypecheckState)))
-
-runElabM ::
-  Swaps ->
-  TypecheckState ->
-  ElabM a ->
-  Either TypeError ([Constraint], TypecheckState, a)
-runElabM swaps tcState value =
-  case either' of
-    ((Right a, constraints), newTcState) -> Right (constraints, newTcState, a)
-    ((Left e, _), _) -> Left e
-  where
-    either' =
-      runState
-        (runReaderT (runWriterT (runExceptT value)) swaps)
-        tcState
-
--- run inference, and substitute everything possible
-elabAndSubst ::
-  Map Name MonoType ->
-  Swaps ->
-  Environment ->
-  TcExpr ->
-  Either TypeError (Substitutions, [Constraint], ElabExpr)
-elabAndSubst _typeMap swaps env expr = do
-  let tcAction = do
-        (elabExpr, constraints) <- listen (elab (envWithBuiltInTypes <> env) expr)
-        subs <- solve constraints
-        --        tyExpr' <- typedHolesCheck typeMap subs tyExpr
-        pure (subs, constraints, elabExpr)
-  (_, _, (subs, constraints, tyExpr)) <- runElabM swaps defaultTcState tcAction
-  pure (subs, constraints, applySubst subs tyExpr)
 
 type TcExpr = Expr Variable Annotation
