+++ compiler/test/golden/PrettyPrint/52c96ab3068acc50952dc4bcfa3fd6f4f7fab5b3964e5793e6a6b70d67da2e62.mimsa
@@ -1,0 +1,4 @@
+let map =
+  \f -> \a -> f(a);
+
+map(?flappy)(1)
