+++ compiler/test/golden/PrettyPrint/94176bc46450a1384de63d71b1d40cf54286cce66877aef8f7f9f321027e331c.mimsa
@@ -1,0 +1,1 @@
+if ?missingFn then 1 else 2
