+++ compiler/test/golden/StoreExpr/52c96ab3068acc50952dc4bcfa3fd6f4f7fab5b3964e5793e6a6b70d67da2e62.json
@@ -1,0 +1,1 @@
+{"storeExpression":{"contents":[[],"map",{"contents":[[],"f",{"contents":[[],"a",{"contents":[[],{"contents":[[],"f"],"tag":"MyVar"},{"contents":[[],"a"],"tag":"MyVar"}],"tag":"MyApp"}],"tag":"MyLambda"}],"tag":"MyLambda"},{"contents":[[],{"contents":[[],{"contents":[[],"map"],"tag":"MyVar"},{"contents":[[],"flappy"],"tag":"MyTypedHole"}],"tag":"MyApp"},{"contents":[[],{"contents":1,"tag":"MyInt"}],"tag":"MyLiteral"}],"tag":"MyApp"}],"tag":"MyLet"},"storeBindings":{},"storeTypeBindings":{}}
