+++ compiler/test/golden/StoreExpr/94176bc46450a1384de63d71b1d40cf54286cce66877aef8f7f9f321027e331c.json
@@ -1,0 +1,1 @@
+{"storeExpression":{"contents":[[],{"contents":[[],"missingFn"],"tag":"MyTypedHole"},{"contents":[[],{"contents":1,"tag":"MyInt"}],"tag":"MyLiteral"},{"contents":[[],{"contents":2,"tag":"MyInt"}],"tag":"MyLiteral"}],"tag":"MyIf"},"storeBindings":{},"storeTypeBindings":{}}
